Audio: Add audio command
/audio is a simple command that allows the user to play audio and interact with the bots audio configuration

To save space, bandwith and keep things simpler, the bot fetches exisiting audio files from the telegram servers trough file id's.
This means that the files must first be uploaded to the telegram servers (using /audio push <section> <name>) which will add an entry to the audio configuration.
This entry stores the name and has it's filed_id assigned.

Signed-off-by: CTXz <ctx.xda@gmail.com>