Power: Add poweroff and reboot methods
If master rights are present, a user may decide to reboot and/or even
poweroff the bot from the chat using /reboot and /poweroff

IMPORTANT NOTE:  /reboot will require you to set the EXEC_PATH flag
under the meta section in your bot configuration. This shall be defined
to the path of TimeForBots.sh.

If the user does not own master rights, access to reboot and poweroff
are denied

To enable reboot and poweroff, add them to the SUPPORTED_METHODS under
your bot configuration

Signed-off-by: CTXz <ctx.xda@gmail.com>